### 2021-04-17 #3185 @whedon
[PRE REVIEW]: plutus: An R package to calculate electricity investments and stranded assets from the Global Change Analysis Model (GCAM)
**Submitting author:** @mengqi-z (<a href="http://orcid.org/0000-0001-5385-2758">Mengqi Zhao</a>)
**Repository:** <a href="https://github.com/JGCRI/plutus.git" target ="_blank">https://github.com/JGCRI/plutus.git</a>
**Version:** v0.1.0
