### 2021-04-17 #3185 @kyleniemeyer
Hello @mengqi-z, thanks for your interest in JOSS. Can you specify which branch has the JOSS paper? I see both `joss-paper-branch_plp` and `joss-paper-branch`.
