### 2021-04-17 #3185 @mengqi-z
> Hello @mengqi-z, thanks for your interest in JOSS. Can you specify which branch has the JOSS paper? I see both `joss-paper-branch_plp` and `joss-paper-branch`.

Hi @kyleniemeyer, the latest paper is located in `joss-paper-branch`. Thanks!
