### 2021-04-21 #3185 @kramea
Sure, I would be interested to review this.

On Wed, Apr 21, 2021 at 2:28 PM Kelly L. Rowland ***@***.***>
wrote:

> Hello @neerajdhanraj <https://github.com/neerajdhanraj> @jordanperr
> <https://github.com/jordanperr> @kramea <https://github.com/kramea> -
>
> Would any of you be interested in and able to take on reviewing this JOSS
> submission?
>
