### 2021-04-22 #3185 @neerajdhanraj
Dear Kelly,

I would like to review this article.

Best regards,
Neeraj

On Thu, 22 Apr 2021, 2:58 am Kelly L. Rowland, ***@***.***>
wrote:

> Hello @neerajdhanraj <https://github.com/neerajdhanraj> @jordanperr
> <https://github.com/jordanperr> @kramea <https://github.com/kramea> -
>
> Would any of you be interested in and able to take on reviewing this JOSS
> submission?
>
