### 2021-05-12 #3212 @kellyrowland
Hi @amygalanter and @joeweaver 👋 - would either of you be willing and available to edit this paper for JOSS?
