### 2021-05-13 #3212 @amygalanter
Hi Kelly,

Yes I am interested. What is the timeline for review?

Thank you,
Sincerely,
Amy Galanter

--
Amy Galanter
Hydrologist
U.S. Geological Survey
New Mexico Water Science Center
6700 Edith Blvd NE, Suite B
Albuquerque, NM  87113
***@***.******@***.***>
(505) 610 - 3140 (mobile)

Pronouns<https://www.mypronouns.org/what-and-why>: she/her/hers
