### 2021-05-14 #3212 @joeweaver
Hi Kelly,

I'm reviewing another paper for JOSS right now, but would interested with
the understanding that I can't start this one right away.

On Thu, May 13, 2021, 00:24 Kelly L. Rowland ***@***.***>
wrote:

> Hi @amygalanter <https://github.com/amygalanter> and @joeweaver
> <https://github.com/joeweaver> 👋 - would either of you be willing and
> available to review this paper for JOSS?
>
