### 2021-05-18 a138b24
Updating github actions for gcc error
