+++ .github/workflows/pkgdown.yaml
@@ -35,6 +35,7 @@ jobs:
       - name: Brew and macOS config
         if: runner.os == 'macOS'
         run: |
+          brew link --overwrite gcc
           brew install pkg-config \
             imagemagick@6 \
             gdal
