+++ .github/workflows/pkgdown.yaml
@@ -31,11 +31,12 @@ jobs:
           path: ${{ env.R_LIBS_USER }}
           key: ${{ runner.os }}-${{ hashFiles('.github/R-version') }}-1-${{ hashFiles('.github/depends.Rds') }}
           restore-keys: ${{ runner.os }}-${{ hashFiles('.github/R-version') }}-1-
-
+        
       - name: Brew and macOS config
         if: runner.os == 'macOS'
         run: |
-          brew link --overwrite gcc
+          brew unlink gcc@8
+          brew unlink gcc@9
           brew install pkg-config \
             imagemagick@6 \
             gdal
