+++ .github/workflows/test_coverage.yml
@@ -18,7 +18,6 @@ jobs:
       - name: Brew and macOS config
         if: runner.os == 'macOS'
         run: |
-          brew link --overwrite gcc
           brew install pkg-config \
             imagemagick@6 \
             gdal
