+++ .github/workflows/pkgdown.yaml
@@ -32,14 +32,6 @@ jobs:
           key: ${{ runner.os }}-${{ hashFiles('.github/R-version') }}-1-${{ hashFiles('.github/depends.Rds') }}
           restore-keys: ${{ runner.os }}-${{ hashFiles('.github/R-version') }}-1-
         
-      - name: Brew and macOS config
-        if: runner.os == 'macOS'
-        run: |
-          brew unlink gcc@8
-          brew unlink gcc@9
-          brew install pkg-config \
-            imagemagick@6 \
-            gdal
         
       - name: Install dependencies
         run: |
