+++ .github/workflows/test_coverage.yml
@@ -15,12 +15,6 @@ jobs:
           r-version: ${{ matrix.r }}
       - uses: r-lib/actions/setup-pandoc@master
       - uses: r-lib/actions/setup-tinytex@master
-      - name: Brew and macOS config
-        if: runner.os == 'macOS'
-        run: |
-          brew install pkg-config \
-            imagemagick@6 \
-            gdal
       - name: Install dependencies
         env:
           R_REMOTES_NO_ERRORS_FROM_WARNINGS: true
