+++ pkgdown.yml
@@ -4,5 +4,5 @@ pkgdown_sha: ~
 articles:
   CaseTutorial: CaseTutorial.html
   gcamInvest: gcamInvest.html
-last_built: 2021-05-18T20:02Z
+last_built: 2021-05-18T21:11Z
 
