### 2021-05-19 #3212 @kellyrowland
Hi @amygalanter @joeweaver - thanks for your responses, let me check in on timelines.

Hi @mengqi-z - do you have any time requirements for having this paper published in JOSS? Whatever the situation is is fine, we're just trying to get a sense of expectations for the potential reviewers.
