### 2021-05-19 #3212 @mengqi-z
> Hi @amygalanter @joeweaver - thanks for your responses, let me check in on timelines.
>
> Hi @mengqi-z - do you have any time requirements for having this paper published in JOSS? Whatever the situation is is fine, we're just trying to get a sense of expectations for the potential reviewers.

Hi @kellyrowland, Thank you for asking. We are hoping to get the paper published in JOSS as soon as possible because we want to cite this JOSS paper in another research work that is going to be published soon. Thanks a lot!
