### 2021-06-04 #3212 @amygalanter
> > > Hi @mengqi-z and @kellyrowland , I finished my review. I appreciate the opportunity to review such a well written and useful package. I checked off all of the boxes except for the last 4 of the **Software paper** because I did not see a **statement of need**, **State of the field**, and **References**. I added this as an "issue" in the repo. Good work @mengqi-z . Please let me know if you need anything additional from me.
> > > Sincerely,
> > > Amy Galanter
> >
> >
> > Hi @amygalanter,
> > Thank you very much for the review and positive feedback to our package. For the software paper, I just checked the PDF generated by whedom in this issue, and it has **Statement of Need** and **References** sections. I am not sure why it is not appearing in the version you have. Would you please double check that? Thank you!
> > Best,
> > Mengqi

My apologies @mengqi-z , I see them now.
