### 2021-06-21 #3212 @kellyrowland
Hi @jtmiclat and @abigailsnyder 👋 - would either of you be willing and available to review this paper for JOSS?
