### 2021-06-22 #3212 @abigailsnyder
@kellyrowland
I work with many of the authors of this software package and have coauthored papers with many of them (nothing to do with this specific package and the paper that wants to cite it).
Please let me know the extent this qualifies as a COI. I’m otherwise able to review this work.

Thanks,
Abigail Snyder
