### 2021-06-24 #3212 @kellyrowland
Hello @arbennett and @richardsc 👋 - would either of you be willing and available to review this paper for JOSS?
