### 2021-06-28 #3212 @richardsc
Hi @kellyrowland -- thanks for asking me, but I'm already reviewing another JOSS submission, and I think this one really isn't in a field that I'd be a good reviewer for.
