### 2021-07-02 #3212 @nick-gauthier
Same as above -- I'm not familiar with GCAM, etc, but happy to look at the R side of things.
