### 2021-09-13 74a6f01
Update Case Tutorial
Fix the problem with rgcam. When setting the working directory the same as the directory of the gcam database, the localDBConn may not work.
