+++ pkgdown.yml
@@ -4,5 +4,5 @@ pkgdown_sha: ~
 articles:
   CaseTutorial: CaseTutorial.html
   gcamInvest: gcamInvest.html
-last_built: 2021-05-19T16:10Z
+last_built: 2021-09-13T02:21Z
 
