### 2021-09-13 JGCRI/plutus#13 @samgdotson
Installation Challenges
I attempted installing ``plutus`` on a fresh installation of Ubuntu 20.04 and ran into two obstacles:
1. ``install.packages('devtools')`` failed because it required additional libraries. I resolved the issue using
```bash
sudo apt install build-essential libcurl4-gnutls-dev libxml2-dev libssl-dev
```
2. ``install.packages('JGCRI/plutus')`` failed because it could not find ``xml2`` (i.e. the ``XML`` dependency was failing).  I resolved this by using
```bash
sudo apt-get install libxml2-dev
```
I'm not sure why I had to install ``libxml2-dev`` twice, but I was able to install plutus after that.

This issue can be closed when:
- [x] the ``README.md`` file includes additional guidance for linux users.
