+++ .github/workflows/rcmd.yml
@@ -1,6 +1,9 @@
 name: build
 
-on: [push]
+on:
+  push:
+    branches:
+      - main
 
 jobs:
   R-CMD:
@@ -17,6 +20,7 @@ jobs:
     env:
       R_REMOTES_NO_ERRORS_FROM_WARNINGS: true
       RSPM: ${{ matrix.config.rspm }}
+      GITHUB_PAT: ${{ secrets.GH_PAT_PLUTUS}}
 
     steps:
       - uses: actions/checkout@v2
