+++ .github/workflows/test_coverage.yml
@@ -1,10 +1,18 @@
-on: [push, pull_request]
-
 name: codecov
+
+on:
+  push:
+    branches:
+      - main
 
 jobs:
   build:
     runs-on: macOS-latest
+    env:
+      R_REMOTES_NO_ERRORS_FROM_WARNINGS: true
+      RSPM: ${{ matrix.config.rspm }}
+      GITHUB_PAT: ${{ secrets.GH_PAT_PLUTUS}}
+      
     strategy:
       matrix:
         r: ['3.6']
