+++ .github/workflows/pkgdown.yaml
@@ -1,4 +1,3 @@
-name: docs
 on:
   push:
     branches:
