### 2021-09-13 JGCRI/plutus#13 @mengqi-z
Hi @samgdotson , thank you for finding out the issues of installing `plutus` on Linux. I updated README to include instructions on installing additional libraries on Linux.
