### 2021-09-16 83bf662
Update CaseTutorial.Rmd
