+++ vignettes/CaseTutorial.Rmd
@@ -21,8 +21,14 @@ vignette: >
     
     - R studio (https://www.rstudio.com/)
 
+2. For Linux users, install following libraries:
+
+```
+sudo apt install build-essential libcurl4-gnutls-dev libxml2-dev libssl-dev
+sudo apt-get install libxml2-dev
+```
     
-2. Open R studio:
+3. Open R studio:
 
 ```
 install.packages('devtools')
@@ -45,7 +51,7 @@ This example will give you a step-by-step instruction on using the core function
 #### 1. Download example dataset
 
   - Download GCAM sample XML database at http://doi.org/10.5281/zenodo.4641500
-  - Unzip dataset to your desired location. In this example, the path to the dataset is: E:/plutus_example/IDBNexus_gcam5p3_HadGEM2-ES_rcp8p5
+  - Unzip dataset to your desired location. Make sure the path is the one that holds the dataset (.basex files). In this example, the path to the dataset is: E:/plutus_example/IDBNexus_gcam5p3_HadGEM2-ES_rcp8p5
 
 #### 2. Prepare your own input data and assumptions
 
@@ -81,10 +87,10 @@ Here are the core arguments we used in this example:
 library(plutus)
 library(dplyr)
 
-# Set your working directory based on the example dataset location
+# Set your directory based on the example dataset location
 workdir <- 'E:/plutus_example/'
 
-# provide path to the desired GCAM database folder.
+# provide path to the desired GCAM database folder that holds .basex files.
 path_to_gcamdatabase <- file.path(workdir, 'IDBNexus_gcam5p3_HadGEM2-ES_rcp8p5')
 # Specify the path to data file folder if any data files has been updated
 # path_to_gcamdataFile <- 'E:/plutus_example/gcamdataFile'
