### 2021-09-17 JGCRI/plutus#14 @samgdotson
Undisclosed dependency on Metis
The full functionality of ``plutus`` depends on installing another package ``metis`` -- which appears to enable plotting. I was able to install ``metis`` and generate plots, however this dependency should be included in the installation steps for ``plutus``.

This issue can be closed when
- [x] the ``README.md`` file either points to ``metis`` installation, or adds instructions to install ``metis``.
