### 2021-09-17 JGCRI/plutus#15 @samgdotson
Instructions for contributing
There are currently no instructions for contributing to the ``plutus`` package. To that end, I offer two examples (however, the file I'm looking for does not need to be as extensive as these):
1. [Github Contrib](https://github.com/github/docs/blob/main/CONTRIBUTING.md)
2. [Pyrk Contrib](https://github.com/pyrk/pyrk/blob/master/CONTRIBUTING.md)

This issue can be closed when
- [x] a ``CONTRIBUTING.md`` file has been added.
