### 2021-09-17 af6fb9b
Update README to include Metis installation guide
