### 2021-09-17 JGCRI/plutus#14 @mengqi-z
Hi @samgdotson , the `README.md` has been updated to include information on `metis` installation guide.
