+++ CONTRIBUTING.md
@@ -1,0 +1,40 @@
+# How to Contribute to `plutus`
+
+Thank you for taking the time to contribute and help us advance the science and architecture of `plutus`. We provide few guidelines that we ask contributors to follow. The guidelines aim to ease the maintainers' organizational and logistical duties, while encouraging development by others.
+
+Before you start:
+
+* Make sure you have a [GitHub account](https://github.com/signup/free).
+* Trivial changes to comments or documentation do not require creating a new issue.
+
+## Did you find a bug?
+
+* Make sure the bug was not already reported in the Github [Issues](https://github.com/JGCRI/plutus/issues).
+* [Open an issue](https://github.com/JGCRI/plutus/issues/new) and clearly describe the issue with as much information as possible. A code sample or an executable test case are recommended.
+  
+## Did you plan to write a patch that fixes a bug?
+
+  * [Open an issue](https://github.com/JGCRI/plutus/issues/new) and clearly describes the problem and discuss how your solution will affect `plutus`.
+  * Fork the repository on GitHub to work on the patch.
+  * Interact with the project maintainers to refine/change/prioritize your issue.
+
+## Making changes
+
+* Start your work on your fork of the repository.
+* Check for unnecessary whitespace with `git diff --check` and format code.
+* Make sure your commit messages are descriptive but succinct, describing what was changed and why, and **reference the relevant issue number**. Make commits of logical units.
+* Make sure you have added the necessary tests for your changes.
+* Run **all** the tests to assure nothing else was accidentally broken.
+
+## Submitting changes
+
+* Submit a pull request with clear documentation of the methodology to the main `plutus` repository.
+* **Your pull request should include one of the following two statements**:
+   * You own the copyright on the code being contributed, and you hereby grant PNNL unlimited license to use this code in this version or any future version of `plutus`. You reserve all other rights to the code.
+   * Somebody else owns the copyright on the code being contributed (e.g., your employer because you did it as part of your work for them); you are authorized by that owner to grant PNNL an unlimited license to use this code in this version or any future version of `plutus`, and you hereby do so. All other rights to the code are reserved by the copyright owner.
+* The core team looks at Pull Requests, and will respond as soon as possible.
+
+# Additional Resources
+
+* [General GitHub documentation](http://help.github.com/)
+* [GitHub pull request documentation](http://help.github.com/send-pull-requests/)
