### 2021-09-18 JGCRI/plutus#15 @mengqi-z
Hi @samgdotson , a `CONTRIBUTING.md` file has been added to the root directory. Thank you for your suggestions!
