### 2021-09-18 #3212 @mengqi-z
> Hi @mengqi-z and @kellyrowland, I completed my review of `plutus`. The functionality promised in the paper exists, I was successfully able to generate plots using this package. All of the boxes have been checked except for two:
>
> 1. The installation box is currently unchecked because there appears to be a dependency on the `metis` package in order to make the plots in your example. Installation instructions for this package should be included or pointed to in the `README.md`.  This can be resolved by closing Issue [Undisclosed dependency on Metis JGCRI/plutus#14](https://github.com/JGCRI/plutus/issues/14).
> 2. The community guidelines box is currently unchecked because I could not find a file with instructions for contributing to `plutus`. This can be resolved by closing Issue [Instructions for contributing JGCRI/plutus#15](https://github.com/JGCRI/plutus/issues/15).
>
> Once addressed, I will complete the review.
>
> Thank you @mengqi-z for adding further installation instructions for Ubuntu users, closing Issue [JGCRI/plutus#13](https://github.com/JGCRI/plutus/issues/13).
>
> Thank you, @kellyrowland for the opportunity to review for JOSS.

Hi @samgdotson , Thank you for your review. I have addressed those two issues. Please check out the updates on `Metis` instruction and CONTRIBUTING.md. We really appreciate the time you took to review `plutus`.
