### 2021-09-20 #3212 @mengqi-z
Hi @kellyrowland , I have published a new release `plutus v1.0.1`, and achieved it in Zenodo DOI:10.5281/zenodo.5517802. Thank you.
