### 2021-09-20 #3212 @kthyng
Hi @mengqi-z! Please alter the metadata in your Zenodo archive so that the title and author list exactly match your JOSS paper.
