### 2021-09-20 #3212 @kthyng
@mengqi-z Your paper looks good! However, the references are missing some capitalization — I noticed the proper noun locations aren't being capitalized. You can preserve capitalization by placing {} around strings in your .bib file. Let me know when you've done this.
