### 2021-09-20 d58f4fd
Update paper.bib to Capitalize the proper none in the reference
