+++ paper/paper.bib
@@ -8,7 +8,7 @@
   number = {4},
   pages = {044026},
   author = {Matthew Binsted and Gokul Iyer and James Edmonds and Adrien Vogt-Schilb and Ricardo Arguello and Angela Cadena and Ricardo Delgado and Felipe Feijoo and Andr{\'{e}} F P Lucena and Haewon McJeon and Fernando Miralles-Wilhelm and Anjali Sharma},
-  title = {Stranded asset implications of the Paris Agreement in Latin America and the Caribbean},
+  title = {Stranded asset implications of the {Paris} {Agreement} in {Latin} {America} and the {Caribbean}},
   journal = {Environmental Research Letters}
 }
 
@@ -61,7 +61,7 @@
   publisher = {Ubiquity Press,  Ltd.},
   volume = {8},
   author = {Zarrar Khan and Thomas Wild and Chris Vernon and Andy Miller and Mohamad Hejazi and Leon Clarke and Fernando Miralles-Wilhelm and Raul Munoz Castillo and Fekadu Moreda and Julia Lacal Bereslawski and Micaela Suriano and Jose Casado},
-  title = {Metis {\textendash} A Tool to Harmonize and Analyze Multi-Sectoral Data and Linkages at Variable Spatial Scales},
+  title = {Metis -- A Tool to Harmonize and Analyze Multi-Sectoral Data and Linkages at Variable Spatial Scales},
   journal = {Journal of Open Research Software}
 }
 
@@ -88,7 +88,7 @@
   number = {7},
   pages = {589--599},
   author = {David L. McCollum and Wenji Zhou and Christoph Bertram and Harmen-Sytze de Boer and Valentina Bosetti and Sebastian Busch and Jacques Despr{\'{e}}s and Laurent Drouet and Johannes Emmerling and Marianne Fay and Oliver Fricko and Shinichiro Fujimori and Matthew Gidden and Mathijs Harmsen and Daniel Huppmann and Gokul Iyer and Volker Krey and Elmar Kriegler and Claire Nicolas and Shonali Pachauri and Simon Parkinson and Miguel Poblete-Cazenave and Peter Rafaj and Narasimha Rao and Julie Rozenberg and Andreas Schmitz and Wolfgang Schoepp and Detlef van Vuuren and Keywan Riahi},
-  title = {Energy investment needs for fulfilling the Paris Agreement and achieving the Sustainable Development Goals},
+  title = {Energy investment needs for fulfilling the {Paris} {Agreement} and achieving the {Sustainable} {Development} {Goals}},
   journal = {Nature Energy}
 }
 
@@ -101,7 +101,7 @@
   volume = {12},
   number = {1},
   author = {Silvia R. Santos da Silva and Mohamad I. Hejazi and Gokul Iyer and Thomas B. Wild and Matthew Binsted and Fernando Miralles-Wilhelm and Pralit Patel and Abigail C. Snyder and Chris R. Vernon},
-  title = {Power sector investment implications of climate impacts on renewable resources in Latin America and the Caribbean},
+  title = {Power sector investment implications of climate impacts on renewable resources in {Latin} {America} and the {Caribbean}},
   journal = {Nature Communications}
 }
 
