### 2021-09-20 #3212 @mengqi-z
> Hi @mengqi-z! Please alter the metadata in your Zenodo archive so that the title and author list exactly match your JOSS paper.

Hi @kthyng , the metadata in the Zenodo archive has been updated to match the title and author list. Thanks for pointing it out.
