### 2024-09-07 #3212 @victoriabenson16
I had my Coinbase wallet wiped out in a single transaction. I didn't know what to do and couldn't get a hold of Coinbase support Teams. This transaction emptied out all my Ethereum and Bitcoin. This was not me and I needed help after losing roughly 821,000 dollars. I went to the police to file a report and they also failed to help me. A gentleman at the stiction walked up to me and handed me a note that contained an Email address (lisaellzabet 662  @  gm ail  .  com) and this number(+1 301 658 0921) he told me this recovery company could help me. And truly after getting in touch them and submitting every documentation within 72 hours i got a mail she was able to recover all my funds and also revealed the men behind the dirty act. I wish to let everyone know that it is possible to get your money back if you contact the right person and follow instructions.
